--- a/src/Streamly/Core.hs
+++ b/src/Streamly/Core.hs
@@ -31,6 +31,7 @@ module Streamly.Core
 
     -- * Composition
     , interleave
+    , roundrobin
 
     -- * Concurrent Stream Vars (SVars)
     , SVar
@@ -271,6 +272,21 @@ interleave m1 m2 = Stream $ \_ stp yld -> do
         yield a Nothing  = yld a (Just m2)
         yield a (Just r) = yld a (Just (interleave m2 r))
     (runStream m1) Nothing stop yield
+
+------------------------------------------------------------------------------
+-- Roundrobin
+------------------------------------------------------------------------------
+
+roundrobin :: Stream m (Stream m a)  -> Stream m a
+roundrobin m = Stream $ \_ stp yld ->
+  let run x = (runStream x) Nothing stp yld
+      yield m1 Nothing = run m1
+      yield m1 (Just rr) = run $ Stream $ \_ stp1 yld1 ->
+        let run1 m2 = (runStream m2) Nothing stp1 yld1
+            yield1 a Nothing  = run1 $ ssing a <> roundrobin rr
+            yield1 a (Just r) = run1 $ ssing a <> roundrobin (rr <> ssing r)
+        in (runStream m1) Nothing stp1 yield1
+  in (runStream m) Nothing stp yield
 
 ------------------------------------------------------------------------------
 -- Spawning threads and collecting result in streamed fashion