--- a/src/Streamly/Streams.hs
+++ b/src/Streamly/Streams.hs
@@ -46,6 +46,7 @@ module Streamly.Streams
 
     -- * Stream Styles
     , StreamT
+    , ReverseT
     , InterleavedT
     , AsyncT
     , ParallelT
@@ -54,6 +55,7 @@ module Streamly.Streams
 
     -- * Type Adapters
     , serially
+    , reversely
     , interleaving
     , asyncly
     , parallely
@@ -63,6 +65,7 @@ module Streamly.Streams
 
     -- * Running Streams
     , runStreamT
+    , runReverseT
     , runInterleavedT
     , runAsyncT
     , runParallelT
@@ -361,6 +364,98 @@ instance (Monad m, Floating a) => Floating (StreamT m a) where
 
     (**)    = liftA2 (**)
     logBase = liftA2 logBase
+
+------------------------------------------------------------------------------
+-- ReverseT
+------------------------------------------------------------------------------
+
+-- TODO: documenation
+
+newtype ReverseT m a = ReverseT {getReverseT :: Stream m a}
+    deriving (Semigroup, Monoid, MonadTrans, MonadIO, MonadThrow)
+
+deriving instance MonadAsync m => Alternative (ReverseT m)
+deriving instance MonadAsync m => MonadPlus (ReverseT m)
+deriving instance (MonadBase b m, Monad m) => MonadBase b (ReverseT m)
+deriving instance MonadError e m => MonadError e (ReverseT m)
+deriving instance MonadReader r m => MonadReader r (ReverseT m)
+deriving instance MonadState s m => MonadState s (ReverseT m)
+
+instance Streaming ReverseT where
+    toStream = getReverseT
+    fromStream = ReverseT
+
+------------------------------------------------------------------------------
+-- Monad
+------------------------------------------------------------------------------
+
+instance Monad m => Monad (ReverseT m) where
+     return = pure
+     s >>= f = ReverseT . roundrobin . getReverseT $ fmap (getReverseT . f) s
+
+------------------------------------------------------------------------------
+-- Applicative
+------------------------------------------------------------------------------
+
+instance Monad m => Applicative (ReverseT m) where
+    pure a = ReverseT $ scons a Nothing
+    (<*>) = ap
+
+------------------------------------------------------------------------------
+-- Functor
+------------------------------------------------------------------------------
+
+instance Monad m => Functor (ReverseT m) where
+    fmap f (ReverseT (Stream m)) = ReverseT $ Stream $ \_ stp yld ->
+        let yield a Nothing  = yld (f a) Nothing
+            yield a (Just r) = yld (f a)
+                               (Just (getReverseT (fmap f (ReverseT r))))
+        in m Nothing stp yield
+
+------------------------------------------------------------------------------
+-- Num
+------------------------------------------------------------------------------
+
+instance (Monad m, Num a) => Num (ReverseT m a) where
+    fromInteger n = pure (fromInteger n)
+
+    negate = fmap negate
+    abs    = fmap abs
+    signum = fmap signum
+
+    (+) = liftA2 (+)
+    (*) = liftA2 (*)
+    (-) = liftA2 (-)
+
+instance (Monad m, Fractional a) => Fractional (ReverseT m a) where
+    fromRational n = pure (fromRational n)
+
+    recip = fmap recip
+
+    (/) = liftA2 (/)
+
+instance (Monad m, Floating a) => Floating (ReverseT m a) where
+    pi = pure pi
+
+    exp  = fmap exp
+    sqrt = fmap sqrt
+    log  = fmap log
+    sin  = fmap sin
+    tan  = fmap tan
+    cos  = fmap cos
+    asin = fmap asin
+    atan = fmap atan
+    acos = fmap acos
+    sinh = fmap sinh
+    tanh = fmap tanh
+    cosh = fmap cosh
+    asinh = fmap asinh
+    atanh = fmap atanh
+    acosh = fmap acosh
+
+    (**)    = liftA2 (**)
+    logBase = liftA2 logBase
+
 
 ------------------------------------------------------------------------------
 -- InterleavedT
@@ -897,6 +992,10 @@ adapt = fromStream . toStream
 serially :: StreamT m a -> StreamT m a
 serially x = x
 
+-- | Interpret an ambiguously typed stream as 'ReverseT'.
+reversely :: ReverseT m a -> ReverseT m a
+reversely x = x
+
 -- | Interpret an ambiguously typed stream as 'InterleavedT'.
 interleaving :: InterleavedT m a -> InterleavedT m a
 interleaving x = x
@@ -924,6 +1023,10 @@ zippingAsync x = x
 -- | Same as @runStreaming . serially@.
 runStreamT :: Monad m => StreamT m a -> m ()
 runStreamT = runStreaming
+
+-- | Same as @runStreaming . reversely@.
+runReverseT :: Monad m => ReverseT m a -> m ()
+runReverseT = runStreaming
 
 -- | Same as @runStreaming . interleaving@.
 runInterleavedT :: Monad m => InterleavedT m a -> m ()