This PR implements #21 !