@harendra-kumar 2018-04-01
LGTM. Just a couple of things:

1) let's call it `CoserialT` and `coserially` as you proposed earlier. The reason is that Coserial at least tells you that it is serial and it is some sort of a variant of serial. `Reverse` does not tell you `reverse` of what, and it does not tell you anything about the serial nature.

2) I find the name `roundrobin` rather confusing, but since it is internal I would not fret about it, you can keep it that way if you prefer. My source of confusion is that if you call it round robin then serial is also round robin because they are not different at all, its just a matter of ordering. For example, If you just swap the two lines in your tests you will get serial.