@harendra-kumar 2018-04-01
Also, did you run the benchmark, to find out the timing in comparison to serial bind?