@harendra-kumar 2018-04-01
One of the reasons why it has limited use, is that reversing the order of the loops can change the logic in most cases in such a way that it may not make sense. My original intention to have this type was for performance reasons where inverting the order of the loops can change perf characteristics without changing the logic. For example, consider this modification to your test case:

```
        it "produce layers reversely" $
          toListReversely (do
            x <- A.each [1, 2]
            y <- A.each [3, 4]
            if (x + y > 5) then mzero else return ()
            z <- A.each [5, 6]
            return (x, y, z))
```

Now, if we run the same code using `serial` then the results will be entirely different, it won't just be ordering change. So it won't be useful. For the use case when we want to just change the order of the loops it may be ok for the programmer to just change it manually. So I  now have a bigger question mark on its utility. Perhaps we should drop the idea altogether?