@harendra-kumar 2018-04-01
Also, one more way of doing things will just confuse the programmer about which one to use, serial or coserial?