@hussein-aitlahcen 2018-04-01
> let's call it CoserialT and coserially as you proposed earlier. The reason is that Coserial at least tells you that it is serial and it is some sort of a variant of serial. Reverse does not tell you reverse of what, and it does not tell you anything about the serial nature.

I am okay with that, especially because it is closely related to the `Serial` stream.

> I find the name roundrobin rather confusing, but since it is internal I would not fret about it, you can keep it that way if you prefer. My source of confusion is that if you call it round robin then serial is also round robin because they are not different at all, its just a matter of ordering. For example, If you just swap the two lines in your tests you will get serial.

Given the joining nature of the function, why not calling it `cojoin` ? I have found that this function is the only difference betwenn `Serial`/`Coserial` in the Monad instance. And actually, the `Serial` counterpart is `join` (which is not yet implemented in the core of Streamly).

Here is an example:
```haskell
sjoin :: Stream m (Stream m a) -> Stream m a
sjoin m = Stream $ \_ stp yld ->
  let run x = (runStream x) Nothing stp yld
      yield m1 Nothing = run m1
      yield m1 (Just rr) = run $ Stream $ \_ stp1 yld1 ->
        let stop = (runStream $ sjoin rr) Nothing stp1 yld1
            yield1 a Nothing  = yld1 a (Just $ sjoin rr)
            yield1 a (Just r) = yld1 a (Just $ sjoin $ scons r (Just rr))
        in (runStream m1) Nothing stop yield1
  in (runStream m) Nothing stp yield

scojoin :: Stream m (Stream m a)  -> Stream m a
scojoin m = Stream $ \_ stp yld ->
  let run x = (runStream x) Nothing stp yld
      yield m1 Nothing = run m1
      yield m1 (Just rr) = run $ Stream $ \_ stp1 yld1 ->
        let stop = (runStream $ scojoin rr) Nothing stp1 yld1
            yield1 a Nothing  = yld1 a (Just $ scojoin rr)
            yield1 a (Just r) = yld1 a (Just $ scojoin (concat rr $ singleton r))
        in (runStream m1) Nothing stop yield1
  in (runStream m) Nothing stp yield
```

Finally, if you think that this will just make the library harder to understand and use, we could make an experimental module, or just drop it :). That was fun to code tho :D