@hussein-aitlahcen 2018-04-01
@harendra-kumar I've just checked the benchmark, we lose something like `30%` of performance in a `CoSerial` execution