@harendra-kumar 2018-04-02
@hussein-aitlahcen I am in favor of dropping it at this point. Don't feel disappointed, I drop a lot more code than this quite often, and quite often we learn from it and it helps in doing something better later. I guess you learnt quite a bit about how streamly works with this. You can pick up some other issues.

I tried benchmarking your `sjoin` implementation for `StreamT` but unfortunately it is quite slower than the current implementation:

```haskell
sjoin :: Stream m (Stream m a) -> Stream m a
sjoin m = Stream $ \_ stp yld ->
  let run x = (runStream x) Nothing stp yld
      yield m1 Nothing = run m1
      yield m1 (Just rr) = run $ Stream $ \_ stp1 yld1 ->
        let stop = (runStream $ sjoin rr) Nothing stp1 yld1
            yield1 a Nothing  = yld1 a (Just $ sjoin rr)
            yield1 a (Just r) = yld1 a (Just $ sjoin $ scons r (Just rr))
        in (runStream m1) Nothing stop yield1
  in (runStream m) Nothing stp yield

instance Monad m => Monad (StreamT m) where
    return = pure
    s >>= f = StreamT . sjoin . getStreamT $ fmap (getStreamT . f) s
```

Current implementation:
```
benchmarked streamly/serial bind/serial
mean                 11.11 ms   (10.97 ms .. 11.25 ms)
```

With sjoin:

```
benchmarked streamly/serial bind/serial
mean                 19.10 ms   (18.80 ms .. 19.55 ms)
```