@hussein-aitlahcen 2018-04-02
I don't mind dropping it :). Was a good experimentation ! Thanks for your advices, I'll work on differents issues.